Finish refactoring Signature TLO code with the removal of the highlighting and commenting capability on a line by line basis
