+++ crits/core/templates/data_widget.html
@@ -6,5 +6,7 @@
 
 <td class="key">Data</td>
 <td class="longtext">
-    <span class="edit_underline" id="object_data">{{data}}</span><span id="object_data_error"></span>
+    <pre>
+        <span class="edit_underline" id="object_data">{{data}}</span><span id="object_data_error"></span>
+    </pre>
 </td>
