--- crits/signatures/templates/signature_data.html
@@ -1,11 +1,0 @@
-{% load filters %}
-    <div class="content_box">
-        <h3 class="titleheader">
-        <span>Data<span>
-        </h3>
-        <div>
-            {% autoescape off %}
-            {{ signature.data|to_line_table }}
-            {% endautoescape %}
-        </div>
-    </div>
